move graphics functions into separate library
